Replace ColorId, HatId, PetId, SkinId with their respective enums

+++ src/Impostor.Api/Innersloth/Customization/SkinType.cs
@@ -1,6 +1,6 @@
 ﻿namespace Impostor.Api.Innersloth.Customization
 {
-    public enum SkinType : byte
+    public enum SkinType : uint
     {
         None = 0,
         Astro = 1,
